Moved NavigationBar to containers
