+++ react-app-typescript/src/containers/NavigationBar/index.css
